+++ ydb/core/kqp/ut/indexes/json/ya.make
@@ -10,6 +10,7 @@ SIZE(MEDIUM)
 SRCS(
     kqp_json_index_corpus.cpp
     kqp_json_index_predicate.cpp
+    kqp_indexes_json_corpus_ut.cpp
     kqp_indexes_json_ut.cpp
 )
